requirements.txt
# RAG Tutorial - Requirements
openai==1.3.0
chromadb==0.4.15
langchain==0.0.350
sentence-transformers==2.2.2
PyPDF2==3.0.1
tiktoken==0.5.1
rank-bm25==0.2.2
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0  # For environment variable management
